refactor(builtins): split methods_narg.rs into cohesive submodules (§7-8)
Split the 3787-line builtins/methods_narg.rs (over the 500-line limit) into a
facade plus ten themed submodules under src/builtins/methods_narg/. All but
one are under 500 lines; dispatch_1arg.rs (1924) holds the single indivisible
~1876-line native_method_1arg dispatch function, which cannot be split at a
function boundary (noted per §7-8 convention).

Submodules and line counts:
- methods_narg.rs (23)      - facade: mod decls + pub(crate) re-exports
- allomorph.rs (100)        - out_of_range_failure + allomorph ACCEPTS helpers
- indent.rs (256)           - str_indent + visual_indent / indent_line helpers
- flatten.rs (53)           - flatten_with_depth / parse_flat_depth / flatten_target
- fmt_contains.rs (128)     - fmt_joinable_target, pair_key_value, contains helpers
- str_match.rs (130)        - is_str_or_match_receiver, native_prefix/suffix/substr_eq
- numeric.rs (196)          - sample_weighted_*_key, int_to_super/subscript, compute_roots
- dispatch_1arg.rs (1924)   - native_method_1arg (single indivisible ~1876-line fn)
- buf.rs (279)              - Buf/Blob read helpers (is_buf_like, resolve_buf_*, read_*)
- dispatch_2arg.rs (418)    - native_method_2arg
- base.rs (394)             - parse_radix_checked, rat_to_base, f64_to_rat, range_pick_n_fast

External callers preserved without edits:
- src/builtins/mod.rs: native_method_1arg, native_method_2arg, native_contains_with_options,
  native_prefix_suffix_with_options, native_substr_eq_with_options
- src/builtins/functions.rs: compute_roots (via methods_narg::compute_roots)
- src/builtins/methods_0arg/dispatch_core_str.rs: fmt_joinable_target

Pure relocation. fn multiset byte-identical. Body super:: refs to builtins siblings
rewritten to absolute crate::builtins:: paths (substr, split, comb, methods_0arg).
Submodule fns made pub(crate) as needed for cross-submodule use; facade re-exports
at original visibility. cargo build + cargo clippy -D warnings + cargo fmt + make test
all pass.
